GetJobsService
